Added Commercial to Geotagging wiki.
Airborne Projects released a geotagging
solution which might be useful for some people.
The additions is a brief description of the
board, how it connects to the APM and 3 images.

The spirit of the section is similar to the
one found on http://ardupilot.org/copter/docs/common-frsky-telemetry.html#hardware-solutions