update docs re: auto mode and arming
This commit implements suggestions from @hamishwillee in issue #804.
Specifically, it clarifies that a copter must be armed before engaging
auto mode, and it updates the list of armable modes.

This also removes the phrase "...and the propellers will begin to spin
slowly", because the propellers when armed are still spinning fast
enough to cause injury.

I have elected not to include Guided mode in these changes, even
though it is an armable mode.

Closes #804