Update skyviper.rst
Made fixes to a few typos and extended information for logging. I also opened an issue on the information about DSMX instead of just changing it here.  In my transmitter, this function seems to be disabled because when I press that button on power on, absolutely nothing happens.  The code clearly indicates that it is being #ifdef-ed out.