Update a branch with a modified submodule
I propose here the solution I used to keep a branch updated with a modified submodule. It was for me the cleanest solution as the switch between the appropriate submodule is really simple when you move from one branch to another. I would be interested to get a feedback about this solution or some possible improvements (for example to avoid the necessity to do the "sync" command before to update).