RC Receivers Update for FrSky and Spektrum
Done primarily to stop answering forum question on why ardupilot is failing to recognize Spektrum signals from satellite receivers. https://github.com/ArduPilot/ardupilot/issues/13117

Also updated some of the FrSky stuff as the X-Lite is quite popular now. PLEASE add the image of an X-Lite to the library; I'm not sure how to include that in this PR.

Other random grammatical and accuracy fixes.