Lua scripts: Update how to add bindings
I believe the description for adding Lua bindings is quite outdated. Especially, there isn't no need anymore for doing the make thing. Cleaning the distro does do it now, which is IMHO a much simpler approach. I wonder if the first sentence in the 3rd bullet should be removed altogether.

By virtue of that I also believe that the reference to the PR is inapproriate since all one has to do now is to PR the additions to the bindings.desc. I thus removed it.