Wrong information about Loiter_turns radius setting
With regards to COPTER, there was a false statement that the circle radius cannot be set as part of a NAV command. It is possible to set number of turn with param3. I don't know if this was different before, but in current firmware this is possible and tested. Mission planner may not display the name of the param3 correctly, but in QGC it is correct and works too.