Update common-accelerometer-calibration.rst
I've seen videos online showing new 'unboxed' flight controllers being calibrated before installing in a vehicle. I guess this might work ok in a quad, but I'm thinking that it might lead to sub-optimal performance in a plane. I'm suggesting making it clearer that the ideal is to calibrate in the vehicle, especially for a plane.

Update common-accelerometer-calibration.rst

Toned it down as suggested by Henry

Update common-accelerometer-calibration.rst

Update common-accelerometer-calibration.rst