Fix wrong parametrer description for find_serial
The `find_serial` binding has two arguments in C++, the protocol and the instance number. In Lua bindings, the first argument is hardwired to 28, while the second one becomes the first Lua argument. The documentation referred to the first C++ argument instead, which caused a lot of confusion.

Now it seems to reflect the current state better.