Update building-setup-windows11.rst to reflect changes with WSL2 and USBIPD.
The older guide mentioned that it was not possible to upload firmware through the USBIPD tool using WSL2 on windows. I just tried it and it worked fine. The only weird thing is that you have to add the -a flag to the usbipd attach command because that ensures that the board reconnects automatically into WSL.

I changed the description of the guide on the section Uploading to a board using WSL2